+++ imod/mf6/wel.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import textwrap
 import warnings
 from typing import Any, Optional, Tuple, Union
 
@@ -11,7 +12,9 @@ import xarray as xr
 import xugrid as xu
 
 import imod
-from imod.logging import init_log_decorator
+from imod.logging import init_log_decorator, logger
+from imod.logging.logging_decorators import standard_log_decorator
+from imod.logging.loglevel import LogLevel
 from imod.mf6.boundary_condition import (
     BoundaryCondition,
     DisStructuredBoundaryCondition,
@@ -26,6 +29,7 @@ from imod.mf6.write_context import WriteContext
 from imod.prepare import assign_wells
 from imod.prepare.layer import create_layered_top
 from imod.schemata import (
+    AllValueSchema,
     AnyNoDataSchema,
     DTypeSchema,
     EmptyIndexesSchema,
@@ -177,7 +181,11 @@ class Well(BoundaryCondition, IPointDataPackage):
     }
     _write_schemata = {
         "screen_top": [AnyNoDataSchema(), EmptyIndexesSchema()],
-        "screen_bottom": [AnyNoDataSchema(), EmptyIndexesSchema()],
+        "screen_bottom": [
+            AnyNoDataSchema(),
+            EmptyIndexesSchema(),
+            AllValueSchema("<", "screen_top"),
+        ],
         "y": [AnyNoDataSchema(), EmptyIndexesSchema()],
         "x": [AnyNoDataSchema(), EmptyIndexesSchema()],
         "rate": [AnyNoDataSchema(), EmptyIndexesSchema()],
@@ -363,6 +371,11 @@ class Well(BoundaryCondition, IPointDataPackage):
 
         return wells_df
 
+    @standard_log_decorator()
+    def _validate(self, schemata: dict, **kwargs) -> dict[str, list[ValidationError]]:
+        kwargs["screen_top"] = self.dataset["screen_top"]
+        return Package._validate(self, schemata, **kwargs)
+
     def __create_assigned_wells(
         self,
         wells_df: pd.DataFrame,
@@ -403,7 +416,7 @@ class Well(BoundaryCondition, IPointDataPackage):
         """
         Create dataset with all variables (rate, concentration), with a similar shape as the cellids.
         """
-        data_vars = ["rate"]
+        data_vars = ["id", "rate"]
         if "concentration" in wells_assigned.columns:
             data_vars.append("concentration")
 
@@ -537,9 +550,6 @@ class Well(BoundaryCondition, IPointDataPackage):
 
         Parameters
         ----------
-        is_partitioned: bool
-        validate: bool
-            Run validation before converting
         active: {xarry.DataArray, xugrid.UgridDataArray}
             Grid with active cells.
         top: {xarry.DataArray, xugrid.UgridDataArray}
@@ -548,6 +558,11 @@ class Well(BoundaryCondition, IPointDataPackage):
             Grid with bottom of model layers.
         k: {xarry.DataArray, xugrid.UgridDataArray}
             Grid with hydraulic conductivities.
+        validate: bool
+            Run validation before converting
+        is_partitioned: bool
+            Set to true if model has been partitioned
+
         Returns
         -------
         Mf6Wel
@@ -563,11 +578,11 @@ class Well(BoundaryCondition, IPointDataPackage):
         minimum_thickness = self.dataset["minimum_thickness"].item()
 
         wells_df = self.__create_wells_df()
+        nwells_df = len(wells_df["id"].unique())
         wells_assigned = self.__create_assigned_wells(
             wells_df, active, top, bottom, k, minimum_k, minimum_thickness
         )
 
-        nwells_df = len(wells_df["id"].unique())
         nwells_assigned = (
             0 if wells_assigned.empty else len(wells_assigned["id"].unique())
         )
@@ -591,7 +606,30 @@ class Well(BoundaryCondition, IPointDataPackage):
         ds["print_flows"] = self["print_flows"].values[()]
         ds["print_input"] = self["print_input"].values[()]
 
+        filtered_wells = [
+            id for id in wells_df["id"].unique() if id not in ds["id"].values
+        ]
+        if len(filtered_wells) > 0:
+            message = self.to_mf6_package_information(filtered_wells)
+            logger.log(loglevel=LogLevel.WARNING, message=message, additional_depth=2)
+
+        ds = ds.drop_vars("id")
+
         return Mf6Wel(**ds.data_vars)
+
+    def to_mf6_package_information(self, filtered_wells):
+        message = textwrap.dedent(
+            """Some wells were not placed in the MF6 well package. This 
+            can be due to inactive cells or permeability/thickness constraints.\n"""
+        )
+        if len(filtered_wells) < 10:
+            message += "The filtered wells are: \n"
+        else:
+            message += " The first 10 unplaced wells are: \n"
+
+        for i in range(min(10, len(filtered_wells))):
+            message += f" id = {filtered_wells[i]} x = {self.dataset['x'][int(filtered_wells[i])].values[()]}  y = {self.dataset['y'][int(filtered_wells[i])].values[()]} \n"
+        return message
 
     def mask(self, domain: GridDataArray) -> Well:
         """
@@ -608,6 +646,46 @@ class Well(BoundaryCondition, IPointDataPackage):
             "layer", errors="ignore"
         )
         return mask_2D(self, domain_2d)
+
+    @classmethod
+    def from_imod5_data(
+        cls,
+        key: str,
+        imod5_data: dict[str, dict[str, GridDataArray]],
+        minimum_k: float = 0.1,
+        minimum_thickness: float = 1.0,
+    ) -> "Well":
+        df: pd.DataFrame = imod5_data[key]["dataframe"]
+
+        # Groupby unique wells, to get dataframes per time.
+        colnames_group = ["x", "y", "filt_top", "filt_bot", "id"]
+        wel_index, df_groups = zip(*df.groupby(colnames_group))
+        # Unpack wel indices by zipping
+        x, y, filt_top, filt_bot, id = zip(*wel_index)
+        # Convert dataframes all groups to DataArrays
+        da_groups = [
+            xr.DataArray(
+                df_group["rate"], dims=("time"), coords={"time": df_group["time"]}
+            )
+            for df_group in df_groups
+        ]
+        # Assign index coordinates
+        da_groups = [
+            da_group.expand_dims(dim="index").assign_coords(index=[i])
+            for i, da_group in enumerate(da_groups)
+        ]
+        # Concatenate datarrays along index dimension
+        well_rate = xr.concat(da_groups, dim="index")
+
+        return cls(
+            x=np.array(x, dtype=float),
+            y=np.array(y, dtype=float),
+            screen_top=np.array(filt_top, dtype=float),
+            screen_bottom=np.array(filt_bot, dtype=float),
+            rate=well_rate,
+            minimum_k=minimum_k,
+            minimum_thickness=minimum_thickness,
+        )
 
 
 class WellDisStructured(DisStructuredBoundaryCondition):