+++ imod/tests/conftest.py
@@ -1,5 +1,6 @@
 import pytest
 
+from .fixtures.backward_compatibility_fixture import imod5_dataset
 from .fixtures.flow_basic_fixture import (
     basic_dis,
     basic_dis__topsystem,
@@ -18,6 +19,10 @@ from .fixtures.flow_basic_unstructured_fixture import (
 )
 from .fixtures.flow_example_fixture import imodflow_model
 from .fixtures.flow_transport_simulation_fixture import flow_transport_simulation
+from .fixtures.imod5_well_data import (
+    well_duplication_import_prj,
+    well_regular_import_prj,
+)
 from .fixtures.mf6_circle_fixture import (
     circle_model,
     circle_model_evt,