+++ imod/tests/test_typing/test_typing_grid.py
@@ -3,7 +3,9 @@ import xugrid as xu
 
 from imod.typing.grid import (
     enforce_dim_order,
+    is_planar_grid,
     is_spatial_grid,
+    is_transient_data_grid,
     merge_with_dictionary,
     preserve_gridtype,
 )
@@ -71,6 +73,27 @@ def test_enforce_dim_order__unstructured(basic_unstructured_dis):
     assert isinstance(actual, type(ibound))
 
 
+def test_is_planar_grid(basic_dis, basic_unstructured_dis):
+    discretizations = [basic_dis, basic_unstructured_dis]
+    for discr in discretizations:
+        ibound, _, _ = discr
+
+        # layer coordinates is present
+        assert not is_planar_grid(ibound)
+
+        # set layer coordinates as present but empty
+        bottom_layer = ibound.sel(layer=3)
+        assert is_planar_grid(bottom_layer)
+
+        # set layer coordinates as  present and not  empty or 0
+        bottom_layer = bottom_layer.expand_dims({"layer": [9]})
+        assert not is_planar_grid(bottom_layer)
+
+        # set layer coordinates as  present and   0
+        bottom_layer.coords["layer"].values[0] = 0
+        assert is_planar_grid(bottom_layer)
+
+
 def test_is_spatial_grid__structured(basic_dis):
     ibound, _, bottom = basic_dis
     ds = xr.Dataset()
@@ -82,6 +105,25 @@ def test_is_spatial_grid__structured(basic_dis):
     assert is_spatial_grid(ds)
 
 
+def test_is_transient_data_grid(basic_dis, basic_unstructured_dis):
+    discretizations = [basic_dis, basic_unstructured_dis]
+
+    for discr in discretizations:
+        ibound, _, _ = discr
+
+        # no time coordinate
+        assert not is_transient_data_grid(ibound)
+
+        #  time coordinate but with single value
+        ibound = ibound.expand_dims({"time": [1]})
+        assert not is_transient_data_grid(ibound)
+
+        #  time coordinate but with several values
+        ibound, _, _ = discr
+        ibound = ibound.expand_dims({"time": [1, 2]})
+        assert is_transient_data_grid(ibound)
+
+
 def test_is_spatial_grid__unstructured(basic_unstructured_dis):
     ibound, _, bottom = basic_unstructured_dis
     grid = ibound.ugrid.grid
@@ -90,23 +132,7 @@ def test_is_spatial_grid__unstructured(basic_unstructured_dis):
     # to dataset.
     ds["ibound"] = (("layer", "mesh2d_nFaces"), ibound)
     ds["bottom"] = bottom
-    uds = xu.UgridDataset(ds, grid)
-
-    assert is_spatial_grid(ibound)
-    assert not is_spatial_grid(bottom)
-    assert is_spatial_grid(uds)
-
-
-def test_merge_dictionary__structured(basic_dis):
-    ibound, _, bottom = basic_dis
-
-    ds = merge_with_dictionary({"ibound": ibound, "bottom": bottom})
-
-    assert isinstance(ds, xr.Dataset)
-    assert isinstance(ds["ibound"], xr.DataArray)
-    assert isinstance(ds["bottom"], xr.DataArray)
-    assert ds["ibound"].dims == ("layer", "y", "x")
-    assert ds["bottom"].dims == ("layer",)
+    _ = xu.UgridDataset(ds, grid)
 
 
 def test_merge_dictionary__unstructured(basic_unstructured_dis):